Import of LBROCARD/Dackup-0.34 from CPAN.
gitpan-cpan-distribution: Dackup
gitpan-cpan-version:      0.34
gitpan-cpan-path:         LBROCARD/Dackup-0.34.tar.gz
gitpan-cpan-author:       LBROCARD
gitpan-cpan-maturity:     released

+++ lib/Dackup.pm
@@ -13,7 +13,7 @@ use Data::Stream::Bulk::Path::Class;
 use Path::Class;
 use Term::ProgressBar::Simple;
 
-our $VERSION = '0.33';
+our $VERSION = '0.34';
 
 has 'directory' => (
     is       => 'ro',
@@ -36,10 +36,19 @@ has 'delete' => (
     isa      => 'Bool',
     required => 1,
 );
+has 'dry_run' => (
+    is      => 'rw',
+    isa     => 'Bool',
+    default => 0,
+);
+has 'verbose' => (
+    is      => 'rw',
+    isa     => 'Bool',
+    default => 0,
+);
 has 'cache' => (
-    is       => 'rw',
-    isa      => 'Dackup::Cache',
-    required => 0,
+    is  => 'rw',
+    isa => 'Dackup::Cache',
 );
 
 __PACKAGE__->meta->make_immutable;
@@ -57,6 +66,9 @@ sub backup {
     my $self        = shift;
     my $source      = $self->source;
     my $destination = $self->destination;
+    my $delete      = $self->delete;
+    my $dry_run     = $self->dry_run;
+    my $verbose     = $self->verbose;
 
     my $source_entries      = $source->entries;
     my $destination_entries = $destination->entries;
@@ -64,28 +76,41 @@ sub backup {
     my ( $entries_to_update, $entries_to_delete )
         = $self->_calc( $source_entries, $destination_entries );
 
-    warn 'to update ' . scalar(@$entries_to_update);
-
-    my $total;
-    if ( $self->delete ) {
-        warn 'to delete ' . scalar(@$entries_to_delete) if $self->delete;
-        $total = scalar(@$entries_to_update) + scalar(@$entries_to_delete);
-    } else {
-        $total = scalar(@$entries_to_update);
-    }
+    my $total = scalar(@$entries_to_update);
+    $total += scalar(@$entries_to_delete) if $delete;
 
     my $progress = Term::ProgressBar::Simple->new($total);
+    $progress->message(
+        'Updating ' . scalar(@$entries_to_update) . ' files' );
+    $progress->message( 'Deleting ' . scalar(@$entries_to_delete) . ' files' )
+        if $delete;
+
     foreach my $entry (@$entries_to_update) {
-        $destination->update( $source, $entry );
+        if ($verbose) {
+            my $source_name      = $source->name($entry);
+            my $destination_name = $destination->name($entry);
+            $progress->message("$source_name -> $destination_name");
+        }
+
+        $destination->update( $source, $entry ) unless $dry_run;
         $progress++;
     }
 
-    if ( $self->delete ) {
+    if ($delete) {
         foreach my $entry (@$entries_to_delete) {
-            $destination->delete($entry);
+            if ($verbose) {
+                my $name = $destination->name($entry);
+                $progress->message("Deleting $name");
+            }
+            $destination->delete($entry) unless $dry_run;
             $progress++;
         }
     }
+
+    $progress->message( 'Updated ' . scalar(@$entries_to_update) . ' files' );
+    $progress->message( 'Deleted ' . scalar(@$entries_to_delete) . ' files' )
+        if $delete;
+
 }
 
 sub _calc {
@@ -154,6 +179,8 @@ Dackup - Flexible file backup
       source      => $source,
       destination => $destination,
       delete      => 0,
+      dry_run     => 0,
+      verbose     => 1,
   );
   $dackup->backup;
